[bugfix] 누락된 투표수 추가

+++ src/main/java/com/may21/trobl/post/domain/PollOption.java
@@ -36,7 +36,7 @@ public class PollOption {
     this.index = index;
   }
 
-  public long getVoteCount() {
+  public int getVoteCount() {
     return pollVotes == null ? 0 : pollVotes.size();
   }
